Añadido Funcion para procesar floats y devolver enteros en la convolución. Juntado el código de la convolución con el código de la reproducción. Primeras pruebas realizadas demuestran su funcionamiento. Falta pruebas en ARM y a mayor escala. (Arreglado pequeño error)

+++ client/src/lib/convolution/convolution.h
@@ -29,7 +29,7 @@ void convolutionOverlapAdd8(float* partX, size_t lx, float localBuffer[], size_t
 
 
 void convolutionOverlapAddINT(float* partX, size_t lx, float localBuffer[], size_t lb, float* h, size_t lh, int y[]);
-void convolutionOverlapAdd8INT(float* partX, size_t lx, float localBuffer[], size_t lb, float* h, size_t lh, float y[]);
+void convolutionOverlapAdd8INT(float* partX, size_t lx, float localBuffer[], size_t lb, float* h, size_t lh, int y[]);
 
 
 #endif //COMUCINACION_CONVOLUTION_H
